feat(which-key): register bindings for whichkey on whichkey loading
plugins are lazily loaded and so the which key register function is also
only executed on loading the plugin. This leads to bugs, that
keybindings that also load the plugin don't appear in the which key
suggestions. I accumulated the registration of bindings to a single
which-key config and also load and setup legendary in the same process.